metadata fixes and split crypto
