+++ core/src/main/kotlin/net/corda/core/crypto/internal/ProviderMap.kt
@@ -7,14 +7,12 @@ import net.corda.core.crypto.Crypto.decodePublicKey
 import net.corda.core.internal.X509EdDSAEngine
 import net.i2p.crypto.eddsa.EdDSAEngine
 import net.i2p.crypto.eddsa.EdDSASecurityProvider
-import org.apache.commons.lang.SystemUtils
 import org.bouncycastle.asn1.ASN1ObjectIdentifier
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo
 import org.bouncycastle.asn1.x509.SubjectPublicKeyInfo
 import org.bouncycastle.jcajce.provider.util.AsymmetricKeyInfoConverter
 import org.bouncycastle.jce.provider.BouncyCastleProvider
 import org.bouncycastle.pqc.jcajce.provider.BouncyCastlePQCProvider
-import java.security.SecureRandom
 import java.security.Security
 
 internal val cordaSecurityProvider = CordaSecurityProvider().also {
@@ -45,9 +43,3 @@ internal val bouncyCastlePQCProvider = BouncyCastlePQCProvider().apply {
 // i.e. if someone removes a Provider and then he/she adds a new one with the same name.
 // The val is private to avoid any harmful state changes.
 internal val providerMap = listOf(cordaBouncyCastleProvider, cordaSecurityProvider, bouncyCastlePQCProvider).map { it.name to it }.toMap()
-internal val platformSecureRandomFactory: () -> SecureRandom = when {
-    SystemUtils.IS_OS_LINUX -> {
-        { SecureRandom.getInstance("NativePRNGNonBlocking") }
-    }
-    else -> SecureRandom::getInstanceStrong
-}
